+++ mt5/cpp/pytorchcpp/pytorch/pytorchcpp.h
@@ -19,8 +19,8 @@
 
 extern "C" // removes decoration of functions __stdcall would mess up but in x64 it is ignored
 {
-    DLL_EXPORT int  fracdiffapply(double signal[], int size, double output[]);
-    DLL_EXPORT void setfracdiffcoefs(double d, int size);
+    DLL_EXPORT int  fracdiffapply(float signal[], int size, float output[]);
+    DLL_EXPORT void setfracdiffcoefs(float d, int size);
     DLL_EXPORT int sadf(float* signal, float* outsadf, float* lagout, int n, int maxw, int minw, int order, bool drift, float gpumem_gb, bool verbose);
     DLL_EXPORT int sadfd_mt5(double* signal, double* outsadf, double* lagout, int n, int maxw, int minw, int order, bool drift, double gpumem_gb, bool verbose);
 }
